Remove class FileAnalyzeThread
