+++ src/main/java/reposense/analyzer/FileAnalyzer.java
@@ -1,12 +1,18 @@
 package reposense.analyzer;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 import reposense.dataobject.FileInfo;
 import reposense.dataobject.RepoConfiguration;
+import reposense.git.GitBlamer;
+import reposense.util.Constants;
 
 
 public class FileAnalyzer {
@@ -40,8 +46,22 @@ public class FileAnalyzer {
     private static ArrayList<FileInfo> processFiles(RepoConfiguration config, List<String> relativePaths) {
         List<FileInfo> fileInfos = Collections.synchronizedList(new ArrayList<FileInfo>());
         for (String relativePath : relativePaths) {
-            FileAnalyzeThread thread = new FileAnalyzeThread(fileInfos, config, relativePath);
-            thread.run();
+            if (isReused(config.getRepoRoot(), relativePath)) {
+                continue;
+            }
+            FileInfo fileInfo = FileInfoGenerator.generateFileInfo(config.getRepoRoot(), relativePath);
+            GitBlamer.aggregateBlameInfo(fileInfo, config);
+            if (config.isNeedCheckStyle()) {
+                CheckStyleParser.aggregateStyleIssue(fileInfo, config.getRepoRoot());
+            }
+            if (config.isAnnotationOverwrite()) {
+                AnnotatorAnalyzer.aggregateAnnotationAuthorInfo(fileInfo, config);
+            }
+            if (!config.getAuthorList().isEmpty() && fileInfo.isAllAuthorsIgnored(config.getAuthorList())) {
+                continue;
+            }
+            fileInfo.constructAuthorContributionMap();
+            fileInfos.add(fileInfo);
         }
         return new ArrayList<>(fileInfos);
     }
@@ -56,4 +76,18 @@ public class FileAnalyzer {
         return false;
     }
 
+    private static boolean isReused(String repoRoot, String relativePath) {
+        File file = new File(repoRoot + '/' + relativePath);
+        try (BufferedReader br = new BufferedReader(new FileReader(file))) {
+            String firstLine = br.readLine();
+            if (firstLine == null || firstLine.contains(Constants.REUSED_TAG)) {
+                return true;
+            }
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return false;
+    }
 }
